requirements.txt
streamlit
google-generativeai
PyPDF2
python-dotenv
langchain
chromadb
